fix: resolve 7 community-reported bugs (#34 #36 #37 #39 #41 #42 #43)
* fix: widen integer storage from 32-bit to 64-bit (#43)

- AST literal union: int → int64_t
- Scanner: strtol → strtoll for integer parsing
- Schema: sqlite3_bind_int → sqlite3_bind_int64 for node/edge properties
- Merge executor: sqlite3_bind_int → sqlite3_bind_int64
- Transform SQL generation: %d → %lld for integer literals
- Exclude expected-failure repro tests from functional test runner

Closes #43

* fix: size(labels(n)) now returns label count, not string length (#42)

Detect list-returning function calls (labels, keys, nodes, relationships,
collect, range, tail, split, json_keys) as arguments to size() and use
json_array_length() instead of LENGTH().

Closes #42

* fix: DELETE + RETURN COUNT now reports actual delete count (#39)

Synthesize COUNT results from nodes_deleted/relationships_deleted
instead of re-querying the now-empty graph. Falls back to re-query
for non-COUNT RETURN clauses.

Closes #39

* fix: OPTIONAL MATCH now respects label filters on target nodes (#34)

Fold label conditions into the target node's LEFT JOIN ON clause using
EXISTS subquery instead of a separate INNER JOIN that would negate
the LEFT JOIN semantics.

Note: WHERE clause after OPTIONAL MATCH still converts LEFT JOIN to
effective INNER JOIN (separate issue from label filtering).

Partially addresses #34

* fix: UNWIND now accepts parameter references and map property access (#37)

- Add AST_NODE_PARAMETER handling in transform_unwind.c (generates json_each(:param))
- Add json_extract for property access on UNWIND JSON values in transform_expr_ops.c
- Fix SELECT column collision between UNWIND and RETURN clause

Closes #37

* fix: startNode(r).name and endNode(r).name now resolve properties (#41)

Add function call handling in transform_property_access() for startNode()
and endNode(). Generates COALESCE property lookup using the node ID
subquery from the function result.

Closes #41

* fix: MERGE variables now persist through WITH into subsequent clauses (#36)

Add MERGE+WITH clause pipeline handler that splits execution at the WITH
boundary. Pre-WITH MERGE+SET executes first, then post-WITH MATCH+MERGE
runs with the MERGE'd variables carried through via re-resolution.

New execute_merge_with_variables() accepts pre-bound variables from the
pipeline. MATCH+SET and MATCH+MERGE patterns now forbid WITH to prevent
incorrect matching of pipelined queries.

Closes #36

* chore: archive completed bug fix tasks

* fix: add stdint.h for int64_t on non-macOS platforms

cypher_schema.c and the bison-generated grammar header need stdint.h
for int64_t. Add %code requires block in grammar and explicit include
in cypher_schema.c.

---------

+++ src/backend/parser/cypher_ast.c
@@ -968,7 +968,7 @@ cypher_path* make_shortest_path(ast_list *elements, path_type type)
     return path;
 }
 
-cypher_literal* make_integer_literal(int value, int location)
+cypher_literal* make_integer_literal(int64_t value, int location)
 {
     cypher_literal *literal = (cypher_literal*)ast_node_create(AST_NODE_LITERAL, location, sizeof(cypher_literal));
     if (!literal) {
@@ -1404,7 +1404,7 @@ void ast_node_print(ast_node *node, int indent)
                 cypher_literal *literal = (cypher_literal*)node;
                 switch (literal->literal_type) {
                     case LITERAL_INTEGER:
-                        printf(" = %d", literal->value.integer);
+                        printf(" = %lld", (long long)literal->value.integer);
                         break;
                     case LITERAL_DECIMAL:
                         printf(" = %f", literal->value.decimal);
